requirements.txt
python-telegram-bot==12.0.0
requests==2.21.0
urllib3==1.24.1
pymongo==3.7.2
pytz
tzlocal